Dev: refactored token lists to hooks and utils
